split the python script test to a different file, so we wouldn't have random failures due to parallel execution
